% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-auth.R
\name{auth_github}
\alias{auth_github}
\title{Authenticate with GitHub}
\usage{
auth_github(...)
}
\arguments{
\item{...}{\code{[character]}\cr
One or more OAuth scopes for GitHub.}
}
\description{
Creates an OAuth token for the requested scopes.
For each combination of scopes, the token will be obtained only once in each
R session, but it will never be cached on the file system.
In most scenarios, this function will be called implicitly by other functions.
}
