% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamCount.R
\name{bamCount}
\alias{bamCount}
\title{Count reads in bam files}
\usage{
bamCount(bamfiles, q = 0, threads = getOption("threads", 1L))
}
\arguments{
\item{q}{An integer >= 0 specifying the minimum quality of a read for it to be counted.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}

\item{bamFiles}{A character vector of paths to bam files.}
}
\description{
\code{bamCount} is a simple wrapper for \code{samtools view -c} for counting the number of reads in a bam file.
}

