% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamParseLengths.R
\name{bamParseLengths}
\alias{bamParseLengths}
\title{Parse reads in bam files by insert size.}
\usage{
bamParseLengths(bamFiles, breaks, threads = getOption("threads", 1L))
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{breaks}{A numeric vector of read lengths used as break points to parse reads by length. For example, if breaks=c(0,100,200), reads will be parsed into reads of length 0-99 and 100-199. Length of vector must beat least 2.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bam.parselengths} separates reads from a paired-end bam file into separate files based on insert sizes using \code{samtools view} and the system \code{awk}.
}

