% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedHist.R
\name{bedHist}
\alias{bedHist}
\title{Plot a histogram of interval sizes in bed files}
\usage{
bedHist(bedFiles, sample = NULL, threads = getOption("threads", 1L), ...)
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{threads}{A positive integer specifying how many beds to process simultaneously.}
}
\description{
\code{bedHist} plots histograms, density functions, or CDFs of interval sizes in bed files.
}

