% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedSizes.R
\name{bedSizes}
\alias{bedSizes}
\title{Retreive interval sizes in bed files}
\usage{
bedSizes(bedFiles, sample = NULL, threads = getOption("threads", 1L))
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bedSizes} returns a list of vectors containing the sizes for intervals in a set of bed files.
}

