% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgHist.R
\name{bgHist}
\alias{bgHist}
\title{Plot a histogram of scores sizes in bedGraph files}
\usage{
bgHist(bgfiles, threads = getOption("threads", 1L), sample = NULL,
  chrom = NULL, ...)
}
\arguments{
\item{threads}{A positive integer specifying how many bedGraph files to process simultaneously.}

\item{bgFiles}{A character vector of paths to bedgraph files.}
}
\description{
\code{bgHist} plots histograms, density functions, or CDFs of scores in bedGraph files.
}

