% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgScores.R
\name{bgScores}
\alias{bgScores}
\title{Retreive scores in bedGraph files}
\usage{
bgScores(bgFiles, threads = getOption("threads", 1L), sample = NULL,
  chrom = NULL, first = NULL)
}
\arguments{
\item{bgFiles}{A character vector of paths to bedgraph files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bgScores} returns a list of vectors containing the scores for intervals in a set of bedGraph files.
}

