% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomart_id_mapping_human}
\alias{biomart_id_mapping_human}
\title{Human gene ID mapping from biomaRt}
\format{
A data frame (tibble) with 91703 rows and 3 columns:
\describe{
\item{ensembl_gene_id}{Human Ensembl gene IDs}
\item{hgnc_symbol}{Human gene symbols}
\item{entrez_gene_id}{Human NCBI/Entrez gene IDs}
}
}
\source{
\url{https://bioconductor.org/packages/biomaRt/}
}
\usage{
biomart_id_mapping_human
}
\value{
An object of class "tbl_df", "tbl", "data.frame"
}
\description{
Human gene ID mapping from biomaRt
}
\keyword{datasets}
