% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomart_id_mapping_mouse}
\alias{biomart_id_mapping_mouse}
\title{Mouse ID mapping from biomaRt}
\format{
A tibble with 55414 rows and 3 columns:
\describe{
\item{ensembl_gene_id}{Mouse Ensembl gene IDs}
\item{mgi_symbol}{Mouse gene name/symbol}
\item{entrez_gene_id}{Mouse NCBI/Entrez gene IDs}
}
}
\usage{
biomart_id_mapping_mouse
}
\description{
Mouse ID mapping from biomaRt
}
\keyword{datasets}
