% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-util.R
\name{browse_travis}
\alias{browse_travis}
\title{Browse the Travis CI web interface}
\usage{
browse_travis(repo = NULL, endpoint = get_endpoint())
}
\arguments{
\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{endpoint}{\verb{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}
}
\description{
Open the current project or any other package on Travis CI.
}
