% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_qc_plots.R
\name{get_rounded_max}
\alias{get_rounded_max}
\title{Find highest number of reads in a long table}
\usage{
get_rounded_max(x, buffer = 1.1, nearest = 1e+07)
}
\arguments{
\item{x}{Data frame of read information}

\item{buffer}{Multiply the largest value by this factor before rounding.
Defaults to \code{1.1}, i.e. adds a 10\% buffer to the maximum value.}

\item{nearest}{Nearest number to round up to. Defaults to \code{10e6}.}
}
\value{
Numeric; maximum number of total reads or counts
}
\description{
Internal helper which takes a long and tidy data frame
containing read or count numbers derived from MultiQC, and finds the
largest read or count value, then round it up to the nearest ten million.
}
\details{
Must contain the columns "sample" and "n_reads".
}
\examples{
count_table <- tibble(
  sample = rep(paste0("s", seq(1, 5)), each = 2),
  read_type = rep(c("unique", "duplicate"), 5),
  n_reads = rnorm(n = 10, mean = 20e6, sd = 5e6)
)
get_rounded_max(count_table)

}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
