% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-meta.R
\name{github_info}
\alias{github_info}
\alias{github_repo}
\title{Github Information}
\usage{
github_info(path = ".")

github_repo(path = ".", info = github_info(path))
}
\arguments{
\item{path}{directory of the git repository}
}
\description{
Retrieves metadata about a git repository from github.
}
