% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-key.R
\name{github_add_key}
\alias{github_add_key}
\title{Add a deploy key to GitHub}
\usage{
github_add_key(pubkey, title = "travis+tic",
  path = usethis::proj_get(), info = github_info(path),
  gh_token = NULL, quiet = FALSE)
}
\arguments{
\item{pubkey}{\code{[pubkey]}\cr
openssl public key, see \code{\link[openssl:read_pubkey]{openssl::read_pubkey()}}.}

\item{title}{\code{[string]}\cr
The title for the new key, default: "travis+tic".}

\item{path}{\code{[string]}\cr
The path to a GitHub-enabled Git repository (or a subdirectory thereof).}

\item{info}{\code{[list]}\cr
GitHub information for the repository, by default obtained through
\code{\link[=github_info]{github_info()}}.}

\item{gh_token}{\code{[Token2.0]}\cr
GitHub authentication token, by default obtained from \code{\link[=auth_github]{auth_github()}} with
the "public_repo" and (if an organization repo) "write:org" scopes.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}
}
\description{
Adds a public deploy key to an existing GitHub repository.
An existing key of the same name is dropped.
}
\seealso{
\code{\link[=github_create_repo]{github_create_repo()}}

Other GitHub functions: \code{\link{github_create_pat}},
  \code{\link{github_create_repo}},
  \code{\link{github_info}}
}
\concept{GitHub functions}
