% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-pat.R
\name{github_create_pat}
\alias{github_create_pat}
\title{Create a GitHub PAT}
\usage{
github_create_pat(path = ".", repo = github_repo(path))
}
\arguments{
\item{path}{\code{[string]}\cr
The path to a GitHub-enabled Git repository (or a subdirectory thereof).}

\item{repo}{\code{[string|numeric]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.
Alternatively, the Travis CI repo ID, e.g. obtained through \code{travis_repo_id()}.}
}
\description{
Prompts the user to create a personal access token for the GitHub API
related to the current repository, e.g., to work around rate limitations.
First, the user is directed to the website where a PAT can be created.
A suggestion for the PAT's title is copied to the clipboard.
Then, the function waits until the user copies a string that resembles a PAT
(i.e., 40 lowercase hexadecimal digits) to the clipboard.
The clipboard is cleared right after obtaining the PAT.
}
\details{
The relatively cumbersome workflow is due to the fact that
creating a PAT requires basic authentication (username + password).
}
\seealso{
Other GitHub functions: \code{\link{github_add_key}},
  \code{\link{github_create_repo}},
  \code{\link{github_info}}
}
