% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-create-repo.R
\name{github_create_repo}
\alias{github_create_repo}
\title{Create a repository on GitHub}
\usage{
github_create_repo(path = ".", name = NULL, org = NULL, private = FALSE,
  gh_token = NULL, quiet = FALSE)
}
\arguments{
\item{path}{\code{[string]}\cr
The path to the existing Git repository.}

\item{name}{\code{[string]}\cr
The name of the new repository on GitHub, default: basename of the
repo directory.}

\item{org}{\code{[string]}\cr
The organization of the new repository on GitHub, default: user namespace,
no organization.}

\item{private}{\code{[flag]}\cr
Must be \code{FALSE}, creation of private repositories not yet supported.}

\item{gh_token}{\code{[Token2.0]}\cr
GitHub authentication token, by default obtained from \code{\link[=auth_github]{auth_github()}} with
the "public_repo" and (if an organization repo) "write:org" scopes.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}
}
\description{
Creates a GitHub repository for an existing Git repository.
}
\seealso{
Other GitHub functions: \code{\link{github_add_key}},
  \code{\link{github_create_pat}},
  \code{\link{github_info}}
}
