% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_github.R, R/helpers_github_bak.R
\name{github_repo}
\alias{github_repo}
\alias{github_info}
\alias{uses_github}
\alias{get_repo_data}
\title{Github information}
\usage{
github_repo(
  path = usethis::proj_get(),
  info = github_info(path, remote = remote),
  remote = "origin"
)

github_info(path = ".", remote = "origin", .token = NULL)

uses_github(path = usethis::proj_get())

get_repo_data(repo, .token = NULL)
}
\arguments{
\item{path}{\verb{[string]}\cr
The path to a GitHub-enabled Git repository (or a subdirectory thereof).}

\item{info}{\verb{[list]}\cr
GitHub information for the repository, by default obtained through
\code{\link[=github_info]{github_info()}}.}

\item{remote}{\verb{[string]}\cr
The Github remote which should be used. Defaults to "origin".}
}
\description{
\code{github_repo()} returns the true repository name as string.

Retrieves metadata about a Git repository from GitHub.

\code{github_info()} returns a list as obtained from the GET "/repos/:repo" API.
}
\seealso{
Other GitHub functions: 
\code{\link{ssh_key_helpers}}
}
\concept{GitHub functions}
