% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gps_rea_hsa}
\alias{gps_rea_hsa}
\title{Gene Pair Signature (GPS) object for human Reactome data}
\format{
A list with nine elements:
\describe{
\item{origRepo}{A list cotaining the original data input to \code{makeGPS}}
\item{L1}{Level 1 gene pair signatures}
\item{L2}{Level 2 gene pair signatures}
\item{L3}{Level 3 gene pair signatures}
\item{L4}{Level 4 gene pair signatures}
\item{L5}{Level 5 gene pair signatures}
\item{repoName}{Name of the repository used to make the GPS object}
\item{pathwaydescriptions}{A data frame of pathway names and IDs}
\item{call}{The call to \code{makeGPS}}
}
}
\usage{
gps_rea_hsa
}
\description{
Reactome human pathway data, constructed by \code{sigora::makeGPS()} and suitable
for use in pathway enrichment with Sigora.
}
\details{
For more information see Sigora's documentation:
\url{https://cran.r-project.org/package=sigora}
}
\keyword{datasets}
