% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello Function
This functions is a hello word function which aims
at giving some instructions}
\usage{
hello(x)
}
\arguments{
\item{x}{A string}
}
\description{
Hello Function
This functions is a hello word function which aims
at giving some instructions
}
