% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_categories}
\alias{reactome_categories}
\title{Pathway hierarchy from Reactome}
\format{
A data frame with 2553 riws and 4 columns
\describe{
\item{id}{Pathway ID}
\item{description}{Pathway name}
\item{level_1}{Top-level parent term}
\item{level_2}{Second-level hierarchy term}
}
}
\usage{
reactome_categories
}
\description{
Reactome pathway hierarchy data for humans
}
\keyword{datasets}
