% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_categories_human}
\alias{reactome_categories_human}
\title{Human pathway hierarchy from Reactome}
\format{
A data frame (tibble) with 2610 rows and 4 columns
\describe{
\item{id}{Pathway ID}
\item{description}{Pathway name}
\item{level_1}{Top-level parent term}
\item{level_2}{Second-level hierarchy term}
}
}
\usage{
reactome_categories_human
}
\description{
Reactome pathway hierarchy data for humans
}
\keyword{datasets}
