% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.stats.R
\name{sam.stats}
\alias{sam.stats}
\title{Get summary statistics for bam files}
\usage{
sam.stats(bamFiles, genomefile, probes = NULL, expandprobes = 200,
  plot = TRUE, uniquescore = 10, cores = 11)
}
\arguments{
\item{bamFiles}{a character vector of paths to position-sorted bam files}

\item{genomefile}{a string of the path to the genome file, a tab-delimited file with chromosome names in column 1 and chromosome sizes in column 2}

\item{probes}{a string of the path to a bed file containing the locations of specific regions of interest to additionally calculate the overlap of reads to these regions}

\item{expandprobes}{a positive integer that is used to expand the region around intervals in 'probes' in which to calculate overlap of reads}

\item{plot}{boolean indicating if a summary plot is drawn for calculated statistics}

\item{uniquescore}{a positive integer used as the threshold for calling an alignment unique}

\item{cores}{a positive integer specifying the number of bam files to process simultaneously}
}
\description{
Calculates the number of duplicate, unique-aligning, multi-aligning, unaligned, and total reads for a set of position-sorted bam files.
}

