% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigora_database}
\alias{sigora_database}
\title{Sigora's database}
\format{
A data frame (tibble) with 60775 rows and 4 columns
\describe{
\item{pathway_id}{Pathway ID}
\item{ensembl_gene_id}{Ensembl ID for constituent genes}
\item{hgnc_symbol}{HGNC symbol for constituent genes}
\item{description}{Pathway name}
}
}
\usage{
sigora_database
}
\description{
Table of Reactome pathways and their genes used by Sigora for enrichment
}
\keyword{datasets}
