% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-key.R, R/setup.R, R/utils.R
\name{ssh_key_helpers}
\alias{ssh_key_helpers}
\alias{github_add_key}
\alias{check_admin_repo}
\alias{get_role_in_repo}
\alias{get_public_key}
\alias{encode_private_key}
\alias{check_private_key_name}
\title{SSH key helpers}
\usage{
github_add_key(
  pubkey,
  repo = get_repo(remote),
  user = get_user(),
  title = "travis",
  remote = "origin"
)

check_admin_repo(owner, user, repo)

get_role_in_repo(owner, user, repo)

get_public_key(key)

encode_private_key(key)

check_private_key_name(string)
}
\arguments{
\item{pubkey}{The public key of the SSH key pair}

\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{user}{The name of the user account}

\item{title}{The title of the key to add}

\item{remote}{\verb{[string]}\cr
The Github remote which should be used. Defaults to "origin".}

\item{owner}{The owner of the repository}

\item{key}{The SSH key pair object}

\item{string}{String to check}
}
\description{
SSH key helpers
}
\keyword{internal}
