% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_clean_deseq2_result.R
\name{tr_clean_deseq2_result}
\alias{tr_clean_deseq2_result}
\title{tr_clean_deseq2_result}
\usage{
tr_clean_deseq2_result(deseq2_result, padjusted = 0.05, foldchange = 1.5)
}
\arguments{
\item{deseq2_result}{Results object for DE genes. \code{DESeq2::results()}
must have been called with 'tidy = TRUE'.}

\item{padjusted}{Threshold for adjusted p-value, defaults to 0.05}

\item{foldchange}{Threshold for fold change, defaults to 1.5}
}
\value{
Tidy dataframe of filtered DE genes.
}
\description{
Helper function to filter and sort results from DESeq2 to aid in
  identifying differentially expressed genes
}
\examples{
\dontrun{
  DESeq2::results(
    dds,
    name = "treatment_peptide_vs_vehicle",
    tidy = TRUE
  ) \%>\%  tr_clean_deseq2_result(.)
}

}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
