% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_deseq2_results.R
\name{tr_deseq2_results}
\alias{tr_deseq2_results}
\title{tr_deseq2_results}
\usage{
tr_deseq2_results(result, pAdj = 0.05, fc = 1.5)
}
\arguments{
\item{result}{Output from call to DESeq2's results function. Must have been
run with the \code{tidy = TRUE} option.}

\item{pAdj}{Cutoff for adjusted p-value. Defaults to 0.05.}

\item{fc}{Cutoff for the fold change value. Defaults to 1.5.}
}
\value{
A dataframe (tibble) of differentially expressed genes, filtered and
  without rownames (first column contains gene identifier).
}
\description{
This function adds two columns to the results object: "ABSLFC"
  is the absolute value of the default "log2FoldChange" column. "FC" is the
  fold change, calculated from the "log2FoldChange". It also filters the
  results based on absolute fold change and adjusted p-value, and orders by
  adjusted p-value.
}
\seealso{
https://www.github.com/travis-m-blimkie/tRavis
}
