% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_enrichment_wrapper.R
\name{tr_enrichment_wrapper}
\alias{tr_enrichment_wrapper}
\title{Wrapper for pathway enrichment with ReactomePA or sigora}
\usage{
tr_enrichment_wrapper(
  tool,
  input_genes,
  species = "human",
  background = NULL,
  gps_repo,
  lvl
)
}
\arguments{
\item{tool}{Package to use for enrichment tests - either "ReactomePA" or
"Sigora"}

\item{input_genes}{Character vector of input genes. For ReactomePA, these
should be Entrez IDs; for Sigora, Ensembl IDs are recommended.}

\item{species}{Desired species for ReactomePA, either "human" (default) or
"mouse"}

\item{background}{Specified gene universe for ReactomePA; default is NULL, or
can be a character vector of Entrez gene IDs.}

\item{gps_repo}{Pathway data to run Sigora with. Should be either one of the
provided options (e.g. \code{data("reaH")}) or one created by
\code{sigora::makeGPS()}}

\item{lvl}{Level to use for Sigora enrichment; should be 4 for Reactome, or 2
for KEGG}
}
\value{
Data frame of results
}
\description{
Wrapper for pathway enrichment with ReactomePA or sigora
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
