% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_get_sigora_genes.R
\name{tr_get_sigora_genes}
\alias{tr_get_sigora_genes}
\title{tr_get_sigora_genes}
\usage{
tr_get_sigora_genes(sigora_result, de_genes, col, reactome_data)
}
\arguments{
\item{sigora_result}{Data frame containing filtered results from Sigora. Uses
the default column name for Reactome ID, "pathwy.id".}

\item{de_genes}{Data frame containing DE genes or other genes used as input
for \code{sigora::sigora()}.}

\item{col}{Column name corresponding to Ensembl gene IDs.}

\item{reactome_data}{Table of all Reactome pathways/genes, with one gene per
row. Expected column names are "reactome_id" and "ensembl_gene_id".}
}
\value{
Modified Sigora result table with additional columns with all the
  genes belonging to a given pathway, and the genes which are also present in
  the input gene list (e.g. DE genes).
}
\description{
Takes \code{sigora::sigora()} result table and corresponding
  input genes, along with the full Reactome (human) data and returns the
  constituent genes for each pathway. Includes all member genes from a
  pathway (all_genes) as well as those which are also present in the user's
  input gene list (candi_genes).
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
