% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_qc_plots.R
\name{tr_qc_plots}
\alias{tr_qc_plots}
\title{Create plotly figures from MultiQC}
\usage{
tr_qc_plots(directory)
}
\arguments{
\item{directory}{Folder containing all the data files generated by MultiQC,
e.g. "multiqc/multiqc_data/"}
}
\value{
List of \code{plotly} objects, and the underlying data
}
\description{
Creates three plotly figures from MultiQC results, and returns
the data along with the plot objects.
}
\examples{
if (FALSE) tr_qc_plots("multiqc_data")

}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
