% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_sigora_wrapper_H.R
\name{tr_sigora_wrapper_H}
\alias{tr_sigora_wrapper_H}
\title{tr_sigora_wrapper_H}
\usage{
tr_sigora_wrapper_H(query_list, database)
}
\arguments{
\item{query_list}{List of input genes as a character vector.}

\item{database}{Database to use, one of "KEGG" or "Reactome". The
\code{level} argument of \code{sigora} is automatically set based on chosen
database (2 for KEGG, 4 for Reactome).}
}
\value{
A tibble corresponding to \code{sigora}'s "summary_results" object with
the "genes" column.
}
\description{
This simple wrapper function allows us to more easily get the
candidate genes from the \code{sigora} results, as they are automatically
included when using the "saveFile" argument. It writes the results to a
temporary file before loading those same results back in and returning them
to the user. The function uses a quiet and safe version of \code{sigora} (based
on the \code{quiet()} and \code{possibly()} \code{purrr} functions), meaning it will
return NULL instead of an error, such as in the case when there are too few
genes given as input.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
