% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_theme_min.R
\name{tr_theme_min}
\alias{tr_theme_min}
\title{Customized ggplot2 minimal theme}
\usage{
tr_theme_min(base_size = 16, base_family = "", grid = "all")
}
\arguments{
\item{base_size}{Base font size applied to all text in the plot (default 18)}

\item{base_family}{Base font type applies to all text in the plot}

\item{grid}{Character to determines how grid lines should be drawn. Options
are "all", "x", "y", or "none".}
}
\value{
A ggplot2 theme
}
\description{
Custom theme that can be applied to ggplot2 plots. Increases
base font size, sets a white background, grey grid, and black border and
text.
}
\examples{
if (FALSE)
  ggplot(mtcars, aes(factor(cyl), mpg)) + geom_boxplot() + tr_theme_min()

}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
