% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_trunc_neatly.R
\name{tr_trunc_neatly}
\alias{tr_trunc_neatly}
\title{Cleanly trim truncate long strings}
\usage{
tr_trunc_neatly(x, l = 60)
}
\arguments{
\item{x}{Input string to trim to desired length, appending an ellipsis to the
end, and without splitting words}

\item{l}{Desired length at which to trim strings}
}
\value{
Character
}
\description{
Trims the input string to the desired length, appending an
ellipsis to the end, without splitting in the middle of a word.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
