% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_write_linux.R
\name{tr_write_linux}
\alias{tr_write_linux}
\title{tr_write_linux}
\usage{
tr_write_linux(x, file)
}
\arguments{
\item{x}{Object to be saved, typically a character vector.}

\item{file}{Name of file to save object to.}
}
\value{
No explicit return, just writes to file.
}
\description{
Custom function to make it easy to write to files with linux
  line endings, to avoid issues caused by running bash script files with
  Windows line endings on a linux system.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
