\name{Travis-package}
\alias{Travis-package}
\alias{Travis}
\docType{package}
\title{
Travis
}
\description{
Transcriptomic View of Genomic Features
}
\details{
Transcriptomic View of Genomic Features
}
\author{
Jia Meng
Maintainer: Jia Meng <jia.meng@xjtlu.edu.cn> 
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
1+1
}
