% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{travis_pubkey}
\alias{travis_pubkey}
\alias{travis_encrypt}
\title{Encrypt for Travis}
\usage{
travis_pubkey(repo = github_repo())

travis_encrypt(data, repo = github_repo(), pubkey = travis_pubkey(repo))
}
\arguments{
\item{repo}{travis repository in "user/repo" format (similar to
\code{\link[remotes:install_github]{remotes::install_github()}})}

\item{data}{raw or character vector to encrypt}

\item{pubkey}{a key returned by \code{\link[=travis_pubkey]{travis_pubkey()}}}
}
\description{
Travis-CI allows for including small pieces of encrypted data with the
configuration file such as passwords or access tokens.
}
\details{
Travis generates a unique RSA keypair for each repository and publishes
the public key via an API. We can use this to encrypt small pieces of
secret data (such as passwords or tokens) in a way that only the Travis
servers can decipher.
}
\examples{
# Encrypt environment variable
\dontrun{
pubkey <- travis_pubkey("jeroenooms/jsonlite")
travis_encrypt("TOKEN=12345", pubkey)
travis_encrypt("TOKEN=12345", "jeroenooms/jsonlite")
}
}
\references{
\url{http://docs.travis-ci.com/user/encryption-keys/}
}
