% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-caches.R
\name{travis_get_caches}
\alias{travis_get_caches}
\alias{travis_delete_caches}
\title{Get/delete repository cache(s) from Travis CI}
\usage{
travis_get_caches(repo = github_repo(), quiet = FALSE)

travis_delete_caches(repo = github_repo())
}
\arguments{
\item{repo}{\code{[string|numeric]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.
Alternatively, the Travis CI repo ID, e.g. obtained through \code{travis_repo_id()}.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}
}
\description{
Return cache information

\code{travis_delete_caches()} returns the repo ID obtained from \code{travis_repo_info()}.
}
\details{
\code{travis_get_caches()} queries the \code{"/repos/:repo/caches"} API.
}
\seealso{
Other Travis CI functions: \code{\link{travis_repo_info}},
  \code{\link{travis_repos}}, \code{\link{travis_set_pat}}
}
\concept{Travis CI functions}
