% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-lint.R
\name{travis_lint}
\alias{travis_lint}
\title{Linting}
\usage{
travis_lint(file = ".travis.yml", repo = github_repo(),
  token = travis_token(repo), quiet = FALSE)
}
\arguments{
\item{file}{A character string specifying a path to a \samp{.travis.yml} file.}

\item{repo}{\code{[string|numeric]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.
Alternatively, the Travis CI repo ID, e.g. obtained through \code{travis_repo_id()}.}

\item{token}{\code{[Token2.0]}\cr
A Travis CI token obtained from \code{\link[=travis_token]{travis_token()}} or \code{\link[=auth_travis]{auth_travis()}}.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}
}
\value{
A list.
}
\description{
This checks if a \samp{.travis.yml} file is valid, and identifies possible errors.
}
\details{
This function may incorrectly report valid \code{.travis.yml} files as broken,
in particular if \code{language: r} is used (which is the default for R projects).
}
\examples{
\dontrun{
travis_lint()
}
}
