% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-pat.R
\name{travis_set_pat}
\alias{travis_set_pat}
\title{Set the GITHUB_PAT variable on Travis}
\usage{
travis_set_pat(pat = NULL, repo = github_repo(), quiet = FALSE)
}
\arguments{
\item{pat}{\code{[string]}\cr
If set, avoids calling \code{github_create_pat()}}

\item{repo}{\code{[string|numeric]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.
Alternatively, the Travis CI repo ID, e.g. obtained through \code{travis_repo_id()}.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}
}
\description{
Assigns a GitHub PAT to the \code{GITHUB_PAT} private variable on Travis CI
via \code{\link[=travis_set_var]{travis_set_var()}},
among others this works around GitHub API rate limitations.
By default, the PAT is obtained from \code{\link[=github_create_pat]{github_create_pat()}}.
}
\seealso{
Other Travis CI functions: \code{\link{travis_get_caches}},
  \code{\link{travis_repo_info}},
  \code{\link{travis_repos}}
}
\concept{Travis CI functions}
