% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{travisHTTP}
\alias{travisHTTP}
\title{Travis CI HTTP Requests}
\usage{
travisHTTP(verb = "GET", path = "", query = list(), body = "",
  endpoint = ".org", encode = "json")
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \code{GET}.}

\item{path}{A character string with the API endpoint (should begin with a
slash).}

\item{query}{A list specifying any query string arguments to pass to the API.
This is used to pass the API token.}

\item{body}{A named list or character string of what should be passed in the
request. Corresponds to the "-d" argument of the \code{curl} command.}

\item{endpoint}{\code{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}

\item{encode}{Encoding format. See \link[httr:POST]{httr::POST}.}
}
\value{
The JSON response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for
Travis CI.
}
\details{
This is mostly an internal function for executing API requests. In
almost all cases, users do not need to access this directly.
}
