% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.R
\name{use_tic}
\alias{use_tic}
\title{Getting tic testing up and running}
\usage{
use_tic(path = usethis::proj_get(), quiet = FALSE)
}
\arguments{
\item{path}{\code{[string]}\cr
The path to the repo to prepare.}

\item{quiet}{\code{[flag]}\cr
Less verbose output? Default: \code{FALSE}.}
}
\description{
Prepares a repo for building and deploying supported by \pkg{tic}.
}
\details{
The preparation consists of the following steps:
\enumerate{
\item If necessary, create a GitHub repository via \code{\link[=use_github]{use_github()}}
\item Enable Travis via \code{\link[=travis_enable]{travis_enable()}}
\item Create a default \code{.travis.yml} file
(overwrite after confirmation in interactive mode only)
\item Create a default \code{appveyor.yml} file
(depending on repo type, overwrite after confirmation
in interactive mode only)
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
\item Enable deployment (if necessary, depending on repo type)
via \code{\link[=use_travis_deploy]{use_travis_deploy()}}
}

This function is aimed at supporting the most common use cases.
Users who require more control are advised to manually call the individual
functions.
}
