% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_travis_deploy}
\alias{use_travis_deploy}
\title{Use travis deploy}
\usage{
use_travis_deploy(pkg = ".", tasks = NULL)
}
\arguments{
\item{pkg}{Package description, can be path or package name. See
\code{\link{as.package}} for more information.}

\item{tasks}{Vector of functions to call on travis}
}
\description{
Use travis deploy
}

