% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{use_travis_vignettes}
\alias{use_travis_vignettes}
\title{Use travis vignettes}
\usage{
use_travis_vignettes(author_email, pkg = ".")
}
\arguments{
\item{author_email}{Email that will be used for commits on your behalf.}

\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}
}
\description{
Use travis vignettes
}

