% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-auth.R
\name{auth_travis}
\alias{auth_travis}
\title{Authenticate to Travis}
\usage{
auth_travis(endpoint = ".org")
}
\arguments{
\item{endpoint}{\code{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}
}
\description{
Authenticates to Travis and returns a Travis API token
}
