% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-auth.R
\name{auth_travis}
\alias{auth_travis}
\title{Authenticate to Travis}
\usage{
auth_travis(endpoint = ".org")
}
\arguments{
\item{endpoint}{´string´\cr
Which endpoint to use. Defaults to ".org".}
}
\description{
Authenticates to Travis and returns a Travis API token
}
