% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-auth.R
\name{auth_travis}
\alias{auth_travis}
\alias{travis_token}
\title{Authenticate with Travis CI}
\usage{
auth_travis(gh_token = NULL)

travis_token(repo = NULL)
}
\arguments{
\item{gh_token}{\code{[Token2.0]}\cr
A GitHub token, by default obtained from \code{\link[=auth_github]{auth_github()}} using the
"read:org", "user:email", "repo_deployment", "repo:status",
"read:repo_hook", and "write:repo_hook" scopes.}

\item{repo}{\code{[string]}\cr
The GitHub repo slug in the format "<user|org>/<repo>", if \code{NULL} no sync
is performed.}
}
\description{
Authenticates with Travis using your Github account. Returns an access token.
The token will be obtained only once in each
R session, but it will never be cached on the file system.
In most scenarios, these functions will be called implicitly by other functions.

\code{auth_travis()} only performs the authentication with Travis CI.

\code{travis_token()} authenticates and checks if the repository is known to
Travis CI. If not, a GitHub sync via \code{\link[=travis_sync]{travis_sync()}} is performed.
}
