% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomart_id_mapping_human}
\alias{biomart_id_mapping_human}
\title{Human ID mapping from biomaRt}
\format{
A tibble with 68005 rows and 3 columns:
\describe{
\item{ensembl_gene_id}{Human Ensembl gene IDs}
\item{hgnc_symbol}{Human gene symbols}
\item{entrez_gene_id}{Human NCBI/Entrez gene IDs}
}
}
\usage{
biomart_id_mapping_human
}
\description{
Human ID mapping from biomaRt
}
\keyword{datasets}
