% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-key.R, R/github-meta.R
\name{github_add_key}
\alias{github_add_key}
\alias{github_info}
\alias{github_repo}
\title{Github Information}
\usage{
github_add_key(pubkey, title = "travis+tic", path = ".",
  info = github_info(path), repo = github_repo(info = info),
  gh_token = NULL, quiet = FALSE)

github_info(path = ".")

github_repo(path = ".", info = github_info(path))
}
\arguments{
\item{pubkey}{openssl public key, see \code{\link[openssl:read_pubkey]{openssl::read_pubkey()}}.}

\item{path}{directory of the git repository}

\item{gh_token}{GitHub token, as returned from \code{\link[=auth_github]{auth_github()}}}
}
\description{
Retrieves metadata about a git repository from github.
}
