% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_info}
\alias{github_info}
\alias{github_repo}
\alias{github_add_key}
\title{Github Information}
\usage{
github_info(path = ".")

github_repo(path = ".", info = github_info(path))

github_add_key(pubkey, path = ".", info = github_info(path),
  repo = github_repo(info = info), gh_token = NULL)
}
\arguments{
\item{path}{directory of the git repository}

\item{pubkey}{openssl public key, see \code{\link[openssl:read_pubkey]{openssl::read_pubkey()}}.}

\item{gh_token}{GitHub token, as returned from \code{\link[=auth_github]{auth_github()}}}
}
\description{
Retrieves metadata about a git repository from github.
}
