% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_info}
\alias{github_info}
\title{Github Information}
\usage{
github_info(path = ".")
}
\arguments{
\item{path}{directory of the git repository}
}
\description{
Retrieves metadata about a git repository from github.
}

