% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-key.R
\name{github_repo}
\alias{github_repo}
\alias{github_info}
\title{Github information}
\usage{
github_repo(path = usethis::proj_get(), info = github_info(path))

github_info(path = usethis::proj_get())
}
\arguments{
\item{path}{\code{[string]}\cr
The path to a GitHub-enabled Git repository (or a subdirectory thereof).}

\item{info}{\code{[list]}\cr
GitHub information for the repository, by default obtained through
\code{\link[=github_info]{github_info()}}.}
}
\description{
\code{github_repo()} returns the true repository name as string.

Retrieves metadata about a Git repository from GitHub.

\code{github_info()} returns a list as obtained from the GET "/repos/:repo" API.
}
\concept{GitHub functions}
