% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigora_database}
\alias{sigora_database}
\title{Sigora's database}
\format{
A data frame (tibble) with 53265 rows and 3 columns
\describe{
\item{pathway_id}{Pathway ID}
\item{entrez_gene_id}{Entrez gene ID for constituent genes}
\item{description}{Pathway name}
}
}
\usage{
sigora_database
}
\description{
Table of Reactome pathways and their genes used by Sigora for enrichment.
}
\keyword{datasets}
