% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_anno_cleaner.R
\name{tr_anno_cleaner}
\alias{tr_anno_cleaner}
\title{Clean annotations from pseudomonas.com}
\usage{
tr_anno_cleaner(input_file, extra_cols = FALSE, fill_names = FALSE)
}
\arguments{
\item{input_file}{Path to the input TSV or CSV file.}

\item{extra_cols}{Logical to determine if start, end, and strand columns
should be included. Defaults to \code{FALSE}.}

\item{fill_names}{Logical to determine if blank/NA genes names should be
filled in with corresponding locus tag. Defaults to \code{FALSE}.}
}
\value{
A data frame (tibble) of the cleaned input file
}
\description{
Given an input CSV or TSV annotation file from
\url{https://pseudomonas.com}, separates and cleans the data, returning a tidy
tibble with the following columns: "locus_tag", "gene_name", and
"product_name". Setting the \code{extra_cols} argument to \code{TRUE} will add the
columns "start", "end", and "strand". Enabling \code{fill_names} will populate
missing gene names with the corresponding locus tag.
}
\examples{
tr_anno_cleaner(
  input_file = paste0(
    "https://pseudomonas.com/downloads/pseudomonas/pgd_r_22_1/",
    "Pseudomonas_aeruginosa_PAO1_107/Pseudomonas_aeruginosa_PAO1_107.csv.gz"
  )
)

}
\references{
Download annotation files from \url{https://pseudomonas.com}
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
