% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_clean_deseq2_result.R
\name{tr_clean_deseq2_result}
\alias{tr_clean_deseq2_result}
\title{Tidy results from DESeq2}
\usage{
tr_clean_deseq2_result(
  deseq2_result,
  p_adjusted = 0.05,
  fold_change = 1.5,
  inform = TRUE
)
}
\arguments{
\item{deseq2_result}{Results object for DE genes, of class \code{DESeqResults}}

\item{p_adjusted}{Threshold for adjusted p-value. Defaults to 0.05.}

\item{fold_change}{Threshold for fold change. Defaults to 1.5.}

\item{inform}{Should a message be printed with the name of the DE comparison
and number of DE genes found? Defaults to \code{TRUE}.}
}
\value{
A data frame (tibble) of filtered DE genes; see \code{?DESeq2::results}
for details on the output.
}
\description{
Helper function to filter and sort results from DESeq2, to aid
in identifying differentially expressed genes.
}
\examples{
ex_deseq_result <-
  readRDS(system.file("extdata", "ex_deseq_results.rds", package = "tRavis"))

tr_clean_deseq2_result(ex_deseq_result)

}
\references{
\url{https://bioconductor.org/packages/DESeq2/}
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
