% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_clean_deseq2_result.R
\name{tr_clean_deseq2_result}
\alias{tr_clean_deseq2_result}
\title{Tidy results from DESeq2}
\usage{
tr_clean_deseq2_result(
  deseq2_result,
  p_adjusted = 0.05,
  fold_change = 1.5,
  inform = TRUE
)
}
\arguments{
\item{deseq2_result}{Results object for DE genes, of class
\code{DESeqResults}}

\item{p_adjusted}{Threshold for adjusted p-value, defaults to 0.05}

\item{fold_change}{Threshold for fold change, defaults to 1.5}

\item{inform}{Should a message be printed with the DE comparison and number
of DE genes found? Defaults to \code{TRUE}.}
}
\value{
Tidy data frame (tibble) of filtered DE genes.
}
\description{
Helper function to filter and sort results from DESeq2 to aid in
identifying differentially expressed genes.
}
\examples{
\dontrun{
  tr_clean_deseq2_result(
    DESeq2::results(dds, name = "treatment_peptide_vs_vehicle")
  )
}

}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
