% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_compare_lists.R
\name{tr_compare_lists}
\alias{tr_compare_lists}
\title{tr_compare_lists}
\usage{
tr_compare_lists(x, y)
}
\arguments{
\item{x}{First vector to compare.}

\item{y}{Second vector to compare.}
}
\value{
A named list of the common and unique elements of x and y.
}
\description{
Performs \code{intersect(x, y)}, \code{setdiff(x, y)}, and
\code{setdiff(y, x)}. Returns these elements in a named list, with names
"common", "unique_x", and "unique_y".
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
