% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_enrichment_wrapper.R
\name{tr_enrichment_wrapper}
\alias{tr_enrichment_wrapper}
\title{Wrapper for pathway enrichment with ReactomePA or sigora}
\usage{
tr_enrichment_wrapper(
  input_genes,
  tool,
  species = "human",
  background = NULL,
  gps_repo = NULL,
  lvl = NULL
)
}
\arguments{
\item{input_genes}{Character vector of input genes. For ReactomePA, these
should be Entrez IDs; for Sigora, Ensembl IDs are recommended.}

\item{tool}{Package to use for enrichment tests - either "ReactomePA" or
"Sigora"}

\item{species}{Desired species for ReactomePA, either "human" (default) or
"mouse". Does not apply to \code{tool = "sigora"}.}

\item{background}{Specified gene universe for ReactomePA; default is NULL, or
can be a character vector of Entrez gene IDs. Does not apply to \code{tool = "sigora"}}

\item{gps_repo}{Gene pair signature object for Sigora to use. Can be one of
the data objects shipped with this package (\code{gps_rea_hsa} or \code{gps_rea_mmu}
for Reactome data for human and mouse, respectively), or one from Sigora
itself - see \code{?sigora::sigora} for details. Does not apply for \code{tool = "ReactomePA"}}

\item{lvl}{Level to use when running Sigora; recommend 4 for Reactome data,
or 2 for KEGG data. Does not apply for \code{tool = "ReactomePA"}}
}
\value{
Data frame (tibble) of results
}
\description{
Wrapper for pathway enrichment with ReactomePA or sigora
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
