% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_genebody_plotly.R
\name{tr_genebody_plotly}
\alias{tr_genebody_plotly}
\title{tr_genebody_plotly}
\usage{
tr_genebody_plotly(qorts_dir)
}
\arguments{
\item{qorts_dir}{Directory containing all the QoRTs results, with each sample
having its own folder. Directory name should include trailing "/".}
}
\value{
Returns a plotly object (html) of the results.
}
\description{
Using functions from the \code{QoRTs} package to read and plot
  gene body coverage data for the upper-middle quartile.
}
\references{
None.
}
\seealso{
https://www.github.com/travis-m-blimkie/tRavis
}
