% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_get_files.R
\name{tr_get_files}
\alias{tr_get_files}
\title{tr_get_files}
\usage{
tr_get_files(folder, pattern = "", date = FALSE, removeString = NULL)
}
\arguments{
\item{folder}{Directory containing files of interest.}

\item{pattern}{Optional pattern to use in file searching.}

\item{date}{Do file names contain a date which should be removed? Should be
of the format "YYYYMMDD"}

\item{removeString}{Optional string which can be removed from file names when
creating names for the list.}
}
\value{
Named list of files to be read.
}
\description{
Function which creates a named list of files in a specified
  directory. Names are trimmed versions of file names, contents of the list
  are the file names themselves. Can be easily piped into
  \code{purrr::map(~read.csv(.))} to create named list of data frames.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
