% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_get_files.R
\name{tr_get_files}
\alias{tr_get_files}
\title{Create a named list of files}
\usage{
tr_get_files(
  folder,
  pattern = "",
  recur = FALSE,
  date = FALSE,
  removeString = NULL
)
}
\arguments{
\item{folder}{Directory containing files of interest.}

\item{pattern}{Optional, case-sensitive pattern to use in file searching. If
no pattern is supplied, all files in the specified directory will be
matched.}

\item{recur}{Whether file listing should be done recursively. Defaults to
FALSE.}

\item{date}{Do file names contain a date which should be removed? Must be
formatted akin to "YYYYMMDD", i.e. all numeric with no spaces, dashes, etc.
Defaults to FALSE.}

\item{removeString}{Optional string which can be removed from file names when
creating names for the output list.}
}
\value{
Named list of files to be read.
}
\description{
Function which creates a named list of files in a specified
directory. The list names are trimmed versions of file names, while
contents of the list are the file names themselves. In this way, it can be
easily piped into \code{purrr::map(~read.csv(.))} to create a named list of
data frames.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
