% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_gtf_cleaner.R
\name{tr_gtf_cleaner}
\alias{tr_gtf_cleaner}
\title{tr_gtf_cleaner}
\usage{
tr_gtf_cleaner(gtf_file)
}
\arguments{
\item{gtf_file}{Path to the input GTF file.}
}
\value{
A dataframe (tibble) of the cleaned GTF file, containing the
  following columns: locus tag, gene name, description, start, end, and
  strand.
}
\description{
Given an input GTF file (bacterial), separates and cleans
  columns, returning a clean and tidy data frame. Only returns locus tag,
  gene name, description, start, end, and strand columns. Only supports PAO1,
  PA14, and LESB58. Uses a single regex to match and extract locus tag for
  all three strains. Will not work with latest GTF version from PGDB (v19).
  Version 18 and below should work fine.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
