% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_qc_plots.R
\name{tr_qc_plots}
\alias{tr_qc_plots}
\title{Create ggplot2 figures from MultiQC results}
\usage{
tr_qc_plots(directory)
}
\arguments{
\item{directory}{Folder containing all the data files generated by MultiQC,
e.g. "multiqc_data/"}
}
\value{
A list with elements "plot" containing the  \code{ggplot} objects, and
"data" containing all the underlying data
}
\description{
Creates four ggplot2 figures from MultiQC results, and returns
the data along with the plot objects. The plots are for FastQC (read counts
and Phred scores), STAR, and HTSeq. See Details for more information.
}
\details{
For the Phred scores, one must open the MultiQC HTML report, and
export the data for "fastqc_per_base_sequence_quality_plot" as a
tab-delimited file (TSV), placing it inside the same directory as the rest.#'
}
\examples{
if (FALSE) tr_qc_plots("multiqc_data")

}
\references{
\url{https://multiqc.info/}
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
