% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_qc_plots.R
\name{tr_qc_plots}
\alias{tr_qc_plots}
\title{tr_qc_plots}
\usage{
tr_qc_plots(directory)
}
\arguments{
\item{directory}{Folder containing all the data files generated by MultiQC}
}
\value{
List of \code{plotly} plot objects
}
\description{
Creates three plotly plots from MultiQC results.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
