% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_qc_plots.R
\name{tr_qc_plots}
\alias{tr_qc_plots}
\title{Create ggplot2 figures from MultiQC results}
\usage{
tr_qc_plots(directory, font_size = 18, threshold_line = 1e+07, limits = NULL)
}
\arguments{
\item{directory}{Folder containing all the data files generated by MultiQC,
e.g. "multiqc_data/"}

\item{font_size}{Base font size (defaults to 18)}

\item{threshold_line}{Provide a number to draw a red dashed line at the
indicated number of reads for FastQC read, STAR, and HTSeq plots. Defaults
to 10e6; set to NULL to disable.}

\item{limits}{Override the upper limit of FastQC read, STAR, and HTSeq plots.
Supply a single number to give all three plots the same limit, or a vector
of three values to modify each individually. Defaults to NULL, which sets
automatic limits.}
}
\value{
A list with elements "plot" containing the \code{ggplot} objects, and
"data" containing all the underlying data
}
\description{
Creates four ggplot2 figures from MultiQC results, and returns
the data along with the plot objects. The plots are for FastQC (read counts
and Phred scores), STAR, and HTSeq. See Details for more information.
}
\details{
For the Phred scores, one must open the MultiQC HTML report, and
export the data for "fastqc_per_base_sequence_quality_plot" as a tab-
delimited file (TSV), placing it inside the same directory as the rest. If
there are too many samples, the data gets saved in
"mqc_fastqc_per_base_sequence_quality_plot_1.txt"; this file is also
checked for automatically.
}
\examples{
if (FALSE) tr_qc_plots("multiqc_data")

}
\references{
\url{https://multiqc.info/}
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
