% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_sort_alphanum.R
\name{tr_sort_alphanum}
\alias{tr_sort_alphanum}
\title{tr_sort_alphanum}
\usage{
tr_sort_alphanum(input_df, sort_col)
}
\arguments{
\item{input_df}{Input data frame or tibble.}

\item{sort_col}{Column to be used in sorting. Can be an index or the name
(quoted).}
}
\value{
Column-sorted data frame
}
\description{
Function to sort a column of alphanumeric strings (e.g. c("a1",
"a11", "a2")) in numeric order (e.g. c("a1", "a2", "a11")). It works with
pipes, and you can provide column name or index as argument \code{sort_col}.
}
\examples{
tr_sort_alphanum(
  input_df = data.frame(colA = c(2, 1, 3), colB = c("m2", "m1", "m3")),
  sort_col = "colB"
)

}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
