% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_test_enrichment.R
\name{tr_test_enrichment}
\alias{tr_test_enrichment}
\title{tr_test_enrichment}
\usage{
tr_test_enrichment(query_genes, enrichment_set, total_genes)
}
\arguments{
\item{query_genes}{List of experimentally or otherwise derived genes, in
which one wishes to test for enrichment.}

\item{enrichment_set}{Set of genes of interest, such as virulence genes.}

\item{total_genes}{Total number of genes for the organism/species.}
}
\value{
Produces the raw p-value from \code{fisher.test()}.
}
\description{
Perfroms Fisher's Exact test to determine enrichment of a set of
  genes of interest compared to some list of experimentally derived genes.
  Assumes \code{alternative = "greater"} in call to \code{fisher.test()}.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
