% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_theme.R
\name{tr_theme}
\alias{tr_theme}
\title{Customized ggplot2 theme}
\usage{
tr_theme(base_size = 18, base_family = "", grid = TRUE)
}
\arguments{
\item{base_size}{Base font size applied to all text in the plot}

\item{base_family}{Base font type applies to all text in the plot}

\item{grid}{Determines if grid lines should be drawn. Defaults to
\code{TRUE}for which only "major" grid lines are drawn.}
}
\value{
A ggplot2 theme
}
\description{
Custom theme that can be applied to ggplot2 plots. Increases
base font size, sets a white background, grey grid, and black border.
}
\examples{
if (FALSE)
  ggplot(mtcars, aes(cyl, mpg)) + geom_point() + tr_theme()

}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/tRavis}
}
