% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-package.R
\docType{package}
\name{travis-package}
\alias{travis}
\alias{travis-package}
\title{travis: Set Up Travis for R package}
\description{
Tools for interacting with the Travis and 'GitHub' APIs for setting up
continuous integration for R packages.
}
\details{
Use \code{\link{github_repo}} to get the name of the current repository
as determined from the \code{origin} remote. The \code{\link{github_info}}
function queries basic information on your repository.
The \code{\link{travis_set_var}} and \code{\link{github_add_key}} functions
interact with the GitHub and Travis configuration of your repository,
and ask for permission the first time you call them.
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

}
