% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-package.R, R/travis.R
\docType{package}
\name{travis-package}
\alias{travis}
\alias{travis-package}
\alias{travis_repo_info}
\alias{travis_repo_id}
\alias{travis_enable}
\alias{travis_set_var}
\alias{travis_delete_var}
\title{travis: Set Up Travis for R package}
\usage{
travis_repo_info(repo = github_repo(), token = travis_token(repo))

travis_repo_id(repo = github_repo(), token = travis_token(repo), ...)

travis_enable(active = TRUE, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token = token))

travis_set_var(name, value, public = FALSE, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token))

travis_delete_var(id, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token))
}
\description{
Tools for interacting with the Travis and 'GitHub' APIs for setting up
continuous integration for R packages.
}
\details{
Use \code{\link{github_repo}} to get the name of the current repository
as determined from the \code{origin} remote. The \code{\link{github_info}}
function queries basic information on your repository.
The \code{\link{travis_set_var}} and \code{\link{github_add_key}} functions
interact with the GitHub and Travis configuration of your repository,
and ask for permission the first time you call them.
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

}
