% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\docType{package}
\name{travis-package}
\alias{travis-package}
\title{Travis CI client package for R}
\description{
Use \code{\link[=github_repo]{github_repo()}} to get the name of the current repository
as determined from the \code{origin} remote.
The following functions simplify integrating R package testing and deployment
with GitHub and Travis CI:
\itemize{
\item \code{\link[=travis_enable]{travis_enable()}} enables Travis CI for your repository,
\item \code{\link[=use_travis_deploy]{use_travis_deploy()}} installs a public deploy key on GitHub and the
corresponding private key on Travis CI to simplify deployments to GitHub
from Travis CI.
}
}
