% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-basic.R, R/travis-enable.R,
%   R/travis-package.R, R/travis-vars.R
\docType{package}
\name{travis_repo_info}
\alias{travis_repo_info}
\alias{travis_repo_id}
\alias{uses_travis}
\alias{travis_enable}
\alias{travis}
\alias{travis-package}
\alias{travis_set_var}
\alias{travis_delete_var}
\title{travis: Set Up Travis for R package}
\usage{
travis_repo_info(repo = github_repo(), token = travis_token(repo))

travis_repo_id(repo = github_repo(), token = travis_token(repo), ...)

uses_travis(repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo = repo, token = token))

travis_enable(active = TRUE, repo = github_repo(),
  token = travis_token(repo = repo), repo_id = travis_repo_id(repo = repo,
  token = token), quiet = FALSE)

travis_set_var(name, value, public = FALSE, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token),
  quiet = FALSE)

travis_delete_var(name, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token), id = travis_get_var_id(name, repo =
  repo, token = token, repo_id = repo_id), quiet = FALSE)
}
\description{
Tools for interacting with the Travis and 'GitHub' APIs for setting up
continuous integration for R packages.
}
\details{
Use \code{\link[=github_repo]{github_repo()}} to get the name of the current repository
as determined from the \code{origin} remote. The \code{\link[=github_info]{github_info()}}
function queries basic information on your repository.
The following functions simplify integrating R package testing and deployment
with GitHub and Travis CI:
\itemize{
\item \code{\link[=github_create_repo]{github_create_repo()}} creates a repository on GitHub,
\item \code{\link[=travis_enable]{travis_enable()}} enables Travis CI for your repository,
\item \code{\link[=travis_set_pat]{travis_set_pat()}} installs a \code{GITHUB_PAT} environment variable on Travis
to work around rate limitations for installing packages from GitHub,
\item \code{\link[=use_travis_deploy]{use_travis_deploy()}} installs a public deploy key on GitHub and the
corresponding private key on Travis CI to simplify deployments to GitHub
from Travis CI.
All these functions ask for permission the first time you call them.
}
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

}
