% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-package.R
\docType{package}
\name{travis-package}
\alias{travis}
\alias{travis-package}
\title{travis: Set Up 'Travis' for Testing and Deployment}
\description{
Tools for interacting with the 'Travis' and 'GitHub' APIs for setting up
continuous integration for R packages and other R-based projects.
}
\details{
Use \code{\link[=github_repo]{github_repo()}} to get the name of the current repository
as determined from the \code{origin} remote. The \code{\link[=github_info]{github_info()}}
function queries basic information on your repository.
The following functions simplify integrating R package testing and deployment
with GitHub and Travis CI:
\itemize{
\item \code{\link[=github_create_repo]{github_create_repo()}} creates a repository on GitHub,
\item \code{\link[=travis_enable]{travis_enable()}} enables Travis CI for your repository,
\item \code{\link[=travis_set_pat]{travis_set_pat()}} installs a \code{GITHUB_PAT} environment variable on Travis
to work around rate limitations for installing packages from GitHub,
\item \code{\link[=use_travis_deploy]{use_travis_deploy()}} installs a public deploy key on GitHub and the
corresponding private key on Travis CI to simplify deployments to GitHub
from Travis CI.
All these functions ask for permission the first time you call them.
}
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (0000-0002-1416-3412)

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
  \item Karthik Ram \email{karthik.ram@gmail.com}
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}
}

Other contributors:
\itemize{
  \item rOpenSci [funder]
}

}
