% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRavis-package.R
\docType{package}
\name{tRavis-package}
\alias{tRavis}
\alias{tRavis-package}
\title{tRavis: Custom functions by Travis Blimkie}
\description{
Includes some custom functions useful for work.
}
\author{
\strong{Maintainer}: Travis Blimkie \email{travis.m.blimkie@gmail.com}

}
\keyword{internal}
