% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{travis_token}
\alias{travis_token}
\alias{travis_set_var}
\title{Authenticate with Travis}
\usage{
travis_token(refresh = FALSE)

travis_set_var(repo_id, name, value, public = FALSE)
}
\description{
Authenticate with Travis using your Github account. Returns an access token.
}

