% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{travis_token}
\alias{travis_token}
\alias{travis_set_var}
\alias{travis_delete_var}
\alias{travis_repo_info}
\alias{travis_repo_id}
\alias{travis_enable}
\title{Authenticate with Travis}
\usage{
travis_token(repo = NULL)

travis_set_var(name, value, public = FALSE, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token))

travis_delete_var(id, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token))

travis_repo_info(repo = github_repo(), token = travis_token(repo))

travis_repo_id(repo = github_repo(), token = travis_token(repo), ...)

travis_enable(active = TRUE, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token = token))
}
\description{
Authenticate with Travis using your Github account. Returns an access token.
}

