% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{travis_token}
\alias{travis_token}
\alias{travis_set_var}
\alias{travis_delete_var}
\alias{travis_repo_info}
\alias{travis_repo_id}
\title{Authenticate with Travis}
\usage{
travis_token()

travis_set_var(name, value, public = FALSE, repo_id = travis_repo_id())

travis_delete_var(id, repo_id = travis_repo_id())

travis_repo_info(repo = github_repo())

travis_repo_id(repo = github_repo())
}
\description{
Authenticate with Travis using your Github account. Returns an access token.
}

