% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-enable.R
\name{travis_enable}
\alias{travis_enable}
\alias{travis_is_enabled}
\title{Activate Travis CI}
\usage{
travis_enable(active = TRUE, repo = github_repo(),
  token = auth_travis())

travis_is_enabled(repo = github_repo(), token = auth_travis())
}
\arguments{
\item{active}{\code{[flag]}\cr
Set to \code{FALSE} to deactivate instead of activating.}

\item{repo}{\code{[string]}\cr
The repository slug to use. Must follow the structure of ´<user>/<repo>´.}

\item{token}{\cr
A Travis CI API token obtained from \code{\link[=auth_travis]{auth_travis()}}.}
}
\description{
Activating Travis CI, and querying activity status.

\code{travis_enable()} activates or deactivates Travis CI for a repo.

\code{travis_is_enabled()} returns if Travis CI is active for a repo.
}
