% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-enable.R
\name{travis_enable}
\alias{travis_enable}
\alias{travis_is_enabled}
\title{Activate Travis CI}
\usage{
travis_enable(active = TRUE, repo = github_repo())

travis_is_enabled(repo = github_repo())
}
\arguments{
\item{active}{\code{[flag]}\cr
Set to \code{FALSE} to deactivate instead of activating.}

\item{repo}{\code{[string|numeric]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.
Alternatively, the Travis CI repo ID, e.g. obtained through \code{travis_repo_id()}.}
}
\description{
Activating Travis CI, and querying activity status.

\code{travis_enable()} activates or deactivates Travis CI for a repo.

\code{travis_is_enabled()} returns if Travis CI is active for a repo.
}
