% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-enable.R
\name{travis_enable}
\alias{travis_enable}
\alias{travis_is_enabled}
\title{Activate Travis CI}
\usage{
travis_enable(
  active = TRUE,
  repo = github_info(remote = remote)$full_name,
  remote = "origin",
  endpoint = get_endpoint()
)

travis_is_enabled(repo = github_repo(), endpoint = get_endpoint())
}
\arguments{
\item{active}{\verb{[flag]}\cr
Set to \code{FALSE} to deactivate instead of activating.}

\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{remote}{\verb{[string]}\cr
The Github remote which should be used. Defaults to "origin".}

\item{endpoint}{\verb{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}
}
\description{
Activating Travis CI, and querying activity status.

\code{travis_enable()} activates or deactivates Travis CI for a repo.

\code{travis_is_enabled()} returns if Travis CI is active for a repo.
}
