% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-builds.R
\name{travis_get_builds}
\alias{travis_get_builds}
\alias{travis_restart_build}
\alias{travis_restart_last_build}
\alias{travis_cancel_build}
\alias{travis_restart_job}
\alias{travis_cancel_job}
\alias{travis_debug_job}
\alias{travis_get_log}
\alias{travis_delete_log}
\title{Travis CI builds}
\usage{
travis_get_builds(repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token))

travis_restart_build(build_id, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token),
  quiet = FALSE)

travis_restart_last_build(repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token), quiet = FALSE)

travis_cancel_build(build_id, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token),
  quiet = FALSE)

travis_restart_job(job_id, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token), quiet = FALSE)

travis_cancel_job(job_id, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token), quiet = FALSE)

travis_debug_job(job_id, log_output = FALSE, repo = github_repo(),
  token = travis_token(repo), repo_id = travis_repo_id(repo, token),
  quiet = FALSE)

travis_get_log(job_id, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token), quiet = FALSE)

travis_delete_log(job_id, repo = github_repo(), token = travis_token(repo),
  repo_id = travis_repo_id(repo, token), quiet = FALSE)
}
\arguments{
\item{repo}{\code{[string]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.}

\item{token}{\code{[Token2.0]}\cr
A Travis CI token obtained from \code{\link[=travis_token]{travis_token()}} or \code{\link[=auth_travis]{auth_travis()}}.}

\item{repo_id}{\code{[string]}\cr
The repo ID, obtained from \code{\link[=travis_repo_id]{travis_repo_id()}} by default.}

\item{build_id}{\code{[integer(1)]}\cr
The build ID, as obtained from \code{travis_get_builds()}.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}

\item{job_id}{\code{[integer(1)]}\cr
The job ID, as obtained from \code{travis_get_builds()}.}

\item{log_output}{Show the debugging output in the publicly visible log? When
set to \code{TRUE}, refrain from issuing commands that might expose secrets.}
}
\description{
Functions around completed or pending Travis CI builds and jobs.

\code{travis_get_builds()} calls the "/builds" API for the current repository.
}
