% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-builds.R
\name{travis_get_builds}
\alias{travis_get_builds}
\alias{travis_restart_build}
\alias{travis_restart_last_build}
\alias{travis_cancel_build}
\alias{travis_get_jobs}
\alias{travis_restart_job}
\alias{travis_cancel_job}
\alias{travis_debug_job}
\alias{travis_get_log}
\alias{travis_delete_log}
\title{Travis CI builds}
\usage{
travis_get_builds(repo = github_repo(), endpoint = get_endpoint())

travis_restart_build(build_id, repo = github_repo(), endpoint = get_endpoint())

travis_restart_last_build(repo = github_repo(), endpoint = get_endpoint())

travis_cancel_build(build_id, repo = github_repo(), endpoint = get_endpoint())

travis_get_jobs(build_id, repo = github_repo(), endpoint = get_endpoint())

travis_restart_job(job_id, repo = github_repo(), endpoint = get_endpoint())

travis_cancel_job(job_id, repo = github_repo(), endpoint = get_endpoint())

travis_debug_job(
  job_id,
  log_output = FALSE,
  repo = github_repo(),
  endpoint = get_endpoint()
)

travis_get_log(job_id, repo = github_repo(), endpoint = get_endpoint())

travis_delete_log(job_id, repo = github_repo(), endpoint = get_endpoint())
}
\arguments{
\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.
\code{travis_get_builds()} calls the "builds" API for the current repository.}

\item{endpoint}{\verb{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}

\item{build_id}{\verb{[integer(1)]}\cr
The build ID, as obtained from \code{travis_get_builds()}.}

\item{job_id}{\verb{[integer(1)]}\cr
The job ID, as obtained from \code{travis_get_builds()}.}

\item{log_output}{Show the debugging output in the publicly visible log? When
set to \code{TRUE}, refrain from issuing commands that might expose secrets.}
}
\description{
Functions around completed or pending Travis CI builds and jobs.
}
