% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-caches.R
\name{travis_get_caches}
\alias{travis_get_caches}
\alias{travis_delete_caches}
\title{Get/delete repository cache(s) from Travis CI}
\usage{
travis_get_caches(repo = github_repo(), endpoint = NULL)

travis_delete_caches(repo = github_repo(), endpoint = NULL)
}
\arguments{
\item{repo}{\code{[string]}\cr
The repository slug to use. Must follow the structure of ´<user>/<repo>´.}
}
\description{
Return cache information

\code{travis_delete_caches()} returns the repo ID obtained from \code{travis_repo_info()}.
}
\details{
\code{travis_get_caches()} queries the \code{"/repos/:repo/caches"} API.
}
\seealso{
Other Travis CI functions: \code{\link{travis_repo_info}},
  \code{\link{travis_repos}}
}
\concept{Travis CI functions}
