% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-caches.R
\name{travis_get_caches}
\alias{travis_get_caches}
\alias{travis_delete_caches}
\title{Get/delete repository cache(s) from Travis CI}
\usage{
travis_get_caches(repo = github_repo(), token = auth_travis())

travis_delete_caches(repo = github_repo(), token = auth_travis())
}
\arguments{
\item{repo}{\code{[string]}\cr
The repository slug to use. Must follow the structure of ´<user>/<repo>´.}

\item{token}{\cr
A Travis CI API token obtained from \code{\link[=auth_travis]{auth_travis()}}.}
}
\description{
Return cache information

\code{travis_get_caches()} queries the "/repos/:repo/caches" API.

\code{travis_delete_caches()} returns the repo ID obtained from \code{travis_repo_info()}.
}
\seealso{
Other Travis CI functions: \code{\link{travis_repo_info}},
  \code{\link{travis_repos}}
}
\concept{Travis CI functions}
