% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-lint.R
\name{travis_lint}
\alias{travis_lint}
\title{Linting}
\usage{
travis_lint(file = ".travis.yml", repo = github_repo(), endpoint = NULL)
}
\arguments{
\item{file}{A character string specifying a path to a \samp{.travis.yml} file.}

\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{endpoint}{\verb{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}
}
\value{
A list.
}
\description{
This checks if a \samp{.travis.yml} file is valid, and identifies possible errors.
}
\details{
This function may incorrectly report valid \code{.travis.yml} files as broken,
in particular if \code{language: r} is used (which is the default for R projects).
}
\examples{
\dontrun{
travis_lint()
}
}
