% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-lint.R
\name{travis_lint}
\alias{travis_lint}
\title{Linting}
\usage{
travis_lint(file = ".travis.yml", repo = github_repo(),
  token = auth_travis())
}
\arguments{
\item{file}{A character string specifying a path to a \samp{.travis.yml} file.}

\item{repo}{\code{[string]}\cr
The repository slug to use. Must follow the structure of ´<user>/<repo>´.}

\item{token}{\cr
A Travis CI API token obtained from \code{\link[=auth_travis]{auth_travis()}}.}
}
\value{
A list.
}
\description{
This checks if a \samp{.travis.yml} file is valid, and identifies possible errors.
}
\details{
This function may incorrectly report valid \code{.travis.yml} files as broken,
in particular if \code{language: r} is used (which is the default for R projects).
}
\examples{
\dontrun{
travis_lint()
}
}
