% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-basic.R
\name{travis_repo_info}
\alias{travis_repo_info}
\alias{travis_has_repo}
\alias{travis_repo_id}
\alias{travis_repo_settings}
\title{Get repository information from Travis CI}
\usage{
travis_repo_info(repo = github_repo(), token = auth_travis())

travis_has_repo(repo = github_repo(), token = auth_travis())

travis_repo_id(repo = github_repo(), token = auth_travis())

travis_repo_settings(repo = github_repo(), token = auth_travis())
}
\arguments{
\item{repo}{\code{[string|numeric]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.
Alternatively, the Travis CI repo ID, e.g. obtained through \code{travis_repo_id()}.}
}
\description{
Return repository information, in particular the repository ID.

\code{travis_repo_info()} queries the "/repos/:repo" API.

\code{travis_repo_id()} returns the repo ID obtained from \code{travis_repo_info()}.

\code{travis_repo_settings()} returns build settings
}
\seealso{
\href{https://docs.travis-ci.com/api}{Travis CI API documentation}

Other Travis CI functions: \code{\link{travis_get_caches}},
  \code{\link{travis_repos}}, \code{\link{travis_set_pat}}
}
\concept{Travis CI functions}
