% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-basic.R
\name{travis_repo_info}
\alias{travis_repo_info}
\alias{travis_repo_id}
\title{Get repository information from Travis CI}
\usage{
travis_repo_info(repo = github_repo(), token = travis_token(repo))

travis_repo_id(repo = github_repo(), token = travis_token(repo))
}
\arguments{
\item{repo}{\code{[string]}\cr
The GitHub repo slug, by default obtained through \code{\link[=github_repo]{github_repo()}}.}

\item{token}{\code{[Token2.0]}\cr
A Travis CI token obtained from \code{\link[=travis_token]{travis_token()}} or \code{\link[=auth_travis]{auth_travis()}}.}
}
\description{
Return repository information, in particular the repository ID.

\code{travis_repo_info()} queries the "/repos/:repo" API.

\code{travis_repo_id()} returns the repo ID obtained from \code{travis_repo_info()}.
}
\seealso{
\href{https://docs.travis-ci.com/api}{Travis CI API documentation}

Other Travis CI functions: \code{\link{travis_accounts}},
  \code{\link{travis_set_pat}}
}
