% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-meta.R
\name{travis_repos}
\alias{travis_repos}
\alias{travis_user}
\title{Retrieve meta information from Travis CI}
\usage{
travis_repos()

travis_user()
}
\description{
Return repositories and user information.

\code{travis_user()} queries the "/users" API.
}
\details{
\code{travis_repositories()} queries the \code{"/repos"} API.
}
\seealso{
\href{https://developer.travis-ci.com/}{Travis CI API documentation}

Other Travis CI functions: \code{\link{travis_get_caches}},
  \code{\link{travis_repo_info}},
  \code{\link{travis_set_pat}}
}
\concept{Travis CI functions}
