% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-util.R
\name{travis_sync}
\alias{travis_sync}
\title{Travis CI utilities}
\usage{
travis_sync(block = TRUE, quiet = FALSE)
}
\arguments{
\item{block}{\code{[flag]}\cr
Set to \code{FALSE} to return immediately instead of waiting.}

\item{quiet}{\code{[flag]}\cr
Set to \code{FALSE} to suppress success message.}
}
\description{
Helper functions for Travis CI.

\code{travis_sync()} initiates synchronization with GitHub and waits for completion
by default.
}
