% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-util.R
\name{travis_sync}
\alias{travis_sync}
\title{Travis CI utilities}
\usage{
travis_sync(block = TRUE, endpoint = get_endpoint())
}
\arguments{
\item{block}{\verb{[flag]}\cr
Set to \code{FALSE} to return immediately instead of waiting.}

\item{endpoint}{\verb{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}
}
\description{
Helper functions for Travis CI.

\code{travis_sync()} initiates synchronization with GitHub and waits for completion
by default.
}
