% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-util.R
\name{travis_sync}
\alias{travis_sync}
\alias{travis_browse}
\title{Travis CI utilities}
\usage{
travis_sync(block = TRUE, token = auth_travis())

travis_browse(repo = github_repo())
}
\arguments{
\item{block}{\code{[flag]}\cr
Set to \code{FALSE} to return immediately instead of waiting.}

\item{token}{\cr
A Travis CI API token obtained from \code{\link[=auth_travis]{auth_travis()}}.}

\item{repo}{\code{[string]}\cr
The repository slug to use. Must follow the structure of ´<user>/<repo>´.}
}
\description{
Helper functions for Travis CI.

\code{travis_sync()} initiates synchronization with GitHub and waits for completion
by default.

\code{travis_browse()} opens a browser pointing to the current repo on  Travis CI.
}
