% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_travis_deploy}
\alias{use_travis_deploy}
\title{Setup deployment for Travis CI}
\usage{
use_travis_deploy(
  path = usethis::proj_get(),
  repo = get_repo_slug(remote),
  key_name_private = NULL,
  key_name_public = NULL,
  endpoint = get_endpoint(),
  remote = "origin",
  quiet = FALSE
)
}
\arguments{
\item{path}{\verb{[string]} \cr
The path to the repository.}

\item{repo}{\verb{[string]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{key_name_private}{\verb{[string]}\cr
The name of the private key of the SSH key pair which will be created.
If not supplied, \code{"TRAVIS_DEPLOY_KEY"} will be used.}

\item{key_name_public}{\verb{[string]}\cr
The name of the private key of the SSH key pair which will be created.
If not supplied, \code{"Deploy key for Travis CI (<endpoint>)"} will be used.}

\item{endpoint}{\verb{[string]}\cr
Which Travis endpoint to use. Defaults to ".org". Accepted values are \code{".com"}
and \code{".org"}. Can be set globally via env var \code{R_TRAVIS}.}

\item{remote}{\verb{[string]}\cr
The Github remote which should be used. Defaults to "origin".}

\item{quiet}{If \code{TRUE}, console output is suppressed.}
}
\description{
Creates a public-private key pair,
adds the public key to the GitHub repository via \code{github_add_key()},
and stores the private key as an encrypted environment variable in Travis CI
via \code{\link[=travis_set_var]{travis_set_var()}},
possibly in a different repository.
The \pkg{tic} companion package contains facilities for installing such a key
during a Travis CI build.
}
