% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_travis_deploy}
\alias{use_travis_deploy}
\title{Setup deployment for Travis CI}
\usage{
use_travis_deploy(path = usethis::proj_get(),
  user = github_user()$owner$login, repo = github_info()$name)
}
\arguments{
\item{path}{\code{[string]} \cr
The path to the repository.}

\item{user}{\code{[string]} \cr
Name of the Github user account.}

\item{repo}{\code{[string]}\cr
The Travis CI repository to add the private key to, default: \code{repo}
(the GitHub repo to which the public deploy key is added).}
}
\description{
Creates a public-private key pair,
adds the public key to the GitHub repository via \code{github_add_key()},
and stores the private key as an encrypted environment variable in Travis CI
via \code{\link[=travis_set_var]{travis_set_var()}},
possibly in a different repository.
The \pkg{tic} companion package contains facilities for installing such a key
during a Travis CI build.
}
